Added Day fourteen part two but still needs optimisation

+++ AdventOfCode2019/Recipes/Element.cs
@@ -6,10 +6,10 @@ namespace AdventOfCode2019.Recipes
 {
     public class Element
     {
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public string Id { get; set; }
 
-        public Element(int amount, string id)
+        public Element(long amount, string id)
         {
             Amount = amount;
             Id = id;
